+++ inc/customizer-controls.php
@@ -387,10 +387,10 @@ class OnePress_Alpha_Color_Control extends WP_Customize_Control {
         <label>
             <?php // Output the label and description if they were passed in.
             if ( isset( $this->label ) && '' !== $this->label ) {
-                echo '<span class="customize-control-title">' . sanitize_text_field( $this->label ) . '</span>';
+                echo '<span class="customize-control-title">' . esc_html( $this->label ) . '</span>';
             }
             if ( isset( $this->description ) && '' !== $this->description ) {
-                echo '<span class="description customize-control-description">' . sanitize_text_field( $this->description ) . '</span>';
+                echo '<span class="description customize-control-description">' . esc_html( $this->description ) . '</span>';
             } ?>
             <input class="alpha-color-control" type="text" data-show-opacity="<?php echo $show_opacity; ?>" data-palette="<?php echo esc_attr( $palette ); ?>" data-default-color="<?php echo esc_attr( $this->settings['default']->default ); ?>" <?php $this->link(); ?>  />
         </label>